requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
numpy==1.14.5
pandas==0.23.2
pycryptodome==3.6.3
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
xlrd==1.1.0